Clean up event topic definitions:
Events prefixed with 'client:' mean that the message _originates_
from the client in the networking module.

Events prefixed with 'server:' mean that the message _originates_
from the server (the node app.js file).

Events prefixed with 'remote:' mean that a message was originated
by another client (not you). This helps us latch the events.

Events with no prefix are originated on the client.